WT-17971 Address review: shared schema-epoch mixin, WT-focused docstring, assertions
- Extract the duplicated schema-epoch/publish helpers from the layered schema
  tests into DisaggSchemaEpochMixin in helper_disagg.py; adopt it in
  test_layered_schema07/10/11/12.
- Reword an unpublished-table docstring to describe WT behavior rather than
  MongoDB multi-document transactions.
- Assert the published vs unpublished table's presence in shared and local
  metadata at the end of the unrelated-writes test.
